requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
exceptiongroup==1.3.1
filelock==3.25.2
Flask==3.1.3
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==26.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
pycparser==3.0
Pygments==2.19.2
pypdfium2==5.6.0
easyocr==1.7.0
python-docx==1.2.0
PyYAML==6.0.3
regex==2026.2.28
rich==14.3.3
safetensors==0.7.0
shellingham==1.5.4
sympy==1.14.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing_extensions==4.15.0
Werkzeug==3.1.6
